webApp/requirement.txt
asgiref==3.8.1
Django==4.2
django-tinymce==4.0.0
sqlparse==0.5.0
typing_extensions==4.11.0
